+++ HISTORY.rst
@@ -48,6 +48,7 @@ Unreleased
 - #716 DataLabel border and fill via ChartFormat
 - fix: #936 text_frame.fit_text crashes if no wrapped representation fits
         in the width of the shape
+- #100 paragraph bullet API
 
 1.0.2 (2024-08-07)
 ++++++++++++++++++